+++ x/recovery/types/errors.go
@@ -1,10 +1,10 @@
 package types
 
 import (
-	sdkerrors "cosmossdk.io/errors"
+	errorsmod "cosmossdk.io/errors"
 )
 
 // errors
 var (
-	ErrBlockedAddress = sdkerrors.Register(ModuleName, 2, "blocked address")
+	ErrBlockedAddress = errorsmod.Register(ModuleName, 2, "blocked address")
 )